Fix compilation issues for 3F2N

+++ sse_mathfun.h
@@ -157,12 +157,12 @@ typedef __m64 v2si;   // vector of 2 int (mmx)
 #define _PI32_CONST(Name, Val)                                            \
   static const ALIGN16_BEG int _pi32_##Name[4] ALIGN16_END = { Val, Val, Val, Val }
 #define _PS_CONST_TYPE(Name, Type, Val)                                 \
-  static const ALIGN16_BEG Type _ps_##Name[4] ALIGN16_END = { Val, Val, Val, Val }
+  static const ALIGN16_BEG Type _ps_##Name[4] ALIGN16_END = { (int)Val, (int)Val, (int)Val, (int)Val }
 
 #define _PD_CONST(Name, Val)                                            \
 	static const ALIGN16_BEG double _pd_##Name[2] ALIGN16_END = { (const double)(Val), (const double)(Val) }
 #define _PD_CONST_TYPE(Name, Type, Val)                                 \
-	static const ALIGN16_BEG Type _pd_##Name[2] ALIGN16_END = { Val, Val }
+	static const ALIGN16_BEG Type _pd_##Name[2] ALIGN16_END = { (long long int)Val, (long long int)Val }
 
 #pragma mark code section
 #ifdef SSE_MATHFUN_WITH_CODE
